requirements.txt
flask
flask-sqlalchemy
flask-login
pymysql
cryptography
pyjwt
google-genai
gunicorn
